refined exception utility functions
