+++ ramannoodle/exceptions.py
@@ -1,5 +1,9 @@
 """Exceptions and warnings for ramannoodle."""
 
+from typing import Any, Sequence
+
+from numpy.typing import NDArray
+
 
 class NoMatchingLineFoundException(Exception):
     """Raised when no line can be found in file."""
@@ -27,3 +31,61 @@ class SymmetryException(Exception):
 
     def __init__(self, reason: str):
         pass
+
+
+def _shape_string(shape: Sequence[int | None]) -> str:
+    """Get a string representing a shape.
+
+    Maps None --> "_", indicating that this element can
+    be anything.
+    """
+    result = "("
+    for i in shape:
+        if i is None:
+            result += "_,"
+        else:
+            result += f"{i},"
+    if len(shape) == 1:
+        return result + ")"
+    return result[:-1] + ")"
+
+
+def get_type_error(name: str, value: Any, correct_type: str) -> TypeError:
+    """Return TypeError for an ndarray argument."""
+    wrong_type = type(value).__name__
+    return TypeError(f"{name} should have type {correct_type}, not {wrong_type}")
+
+
+def verify_ndarray(name: str, array: NDArray) -> None:
+    """Verify type of NDArray .
+
+    We should avoid calling this function wherever possible (EATF)
+    """
+    try:
+        _ = array.shape
+    except AttributeError as exc:
+        raise get_type_error(name, array, "ndarray") from exc
+
+
+def verify_ndarray_shape(
+    name: str, array: NDArray, shape: Sequence[int | None]
+) -> None:
+    """Verify an NDArray's shape.
+
+    We should avoid calling this function whenever possible (EATF).
+
+    Parameters
+    ----------
+    shape
+        int elements will be checked, None elements will not be.
+    """
+    try:
+        if len(shape) != array.ndim:
+            shape_spec = f"{_shape_string(array.shape)} != {_shape_string(shape)}"
+            raise ValueError(f"{name} has wrong shape: {shape_spec}")
+        for d1, d2 in zip(array.shape, shape, strict=True):
+            if d2 is not None and d1 != d2:
+                shape_spec = f"{_shape_string(array.shape)} != {_shape_string(shape)}"
+                raise ValueError(f"{name} has wrong shape: {shape_spec}")
+    except AttributeError as exc:
+        raise get_type_error(name, array, "ndarray") from exc
